refactor: move composables and models
