--- app/src/main/java/com/pereyrarg11/drinks/feature/home/presentation/HomeEvent.kt
@@ -1,5 +1,0 @@
-package com.pereyrarg11.drinks.feature.home.presentation
-
-sealed class HomeEvent {
-    object Refresh : HomeEvent()
-}
