+++ app/src/main/java/com/pereyrarg11/drinks/feature/home/presentation/HomeScreen.kt
@@ -8,6 +8,7 @@ import com.pereyrarg11.drinks.core.presentation.ErrorScreen
 import com.pereyrarg11.drinks.core.presentation.LoadingScreen
 import com.pereyrarg11.drinks.core.presentation.MainScreen
 import com.pereyrarg11.drinks.core.presentation.navigation.ScreenRoute
+import com.pereyrarg11.drinks.feature.home.presentation.components.HomeContent
 
 @Composable
 fun HomeScreen(
@@ -22,7 +23,7 @@ fun HomeScreen(
 
         when {
             state.hasError -> ErrorScreen(message = state.errorMessage)
-            !state.isLoading && !state.hasError -> HomeSuccessScreen(
+            !state.isLoading && !state.hasError -> HomeContent(
                 sections = state.sections,
                 onClickFilter = { model ->
                     viewModel.onClickHomeFilter(model)
