+++ app/src/main/java/com/pereyrarg11/drinks/feature/home/presentation/model/HomeEvent.kt
@@ -1,0 +1,5 @@
+package com.pereyrarg11.drinks.feature.home.presentation.model
+
+sealed class HomeEvent {
+    object Refresh : HomeEvent()
+}
